% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBio.R
\name{PlotBio}
\alias{PlotBio}
\title{Plot Level Biomass Computation}
\usage{
PlotBio(field_data, outfile, PlotIDfile)
}
\arguments{
\item{field_data:}{Dataframe of SpeciesName, StateName, Girth (cm)}

\item{outfile:}{Path to output file where the outputs have to be generated.}

\item{PlotIDfile:}{File containing PlotID and Lat and Long for joining with Biomass outputs.}
}
\description{
This function calculates Biomass values of all trees on PlotID basis.
}
\details{
Biomass is computed on a tree basis using specieswise allometric equations with Diameter > 10cm.
An incremental biomass approach is used for trees of diameter < 10cm.

The function takes minimum of one input to calculate the biomass (field_data)
Additionally it take PlotIDfile to join Lat and Long information from a seperate file based on PlotID column
The output of the function can be directly saved into a csv file using outfile argument (See examples).
}
\examples{
##Sample data is in data folder
field_data = read.csv("data/sample_tree_data.csv") #read a csv file to a dataframe
pidfile = "data/sample_plotid_file.csv" # path to plotid locations file
outputfile = "data/sample_output.csv" # path to output file.

## only using field_data
PlotBio(field_data) # shows the summary of the plot biomass
output = PlotBio(field_data) # can be stored in a data frame

## Using PlotIDfile
output = PlotBio(field_data,PlotIDfile=pidfile) # attaches lat and long information

## save output to a separate file
PlotBio(field_data,outfile = outputfile,PlotIDfile=pidfile)
# output is summarised along with lat and long information

## or
PlotBio(field_data,outfile = outputfile) #no lat and long file.
}
\keyword{Biomass}
\keyword{Plot}

