% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WKTPlotBM.R
\name{WKTPlotBM}
\alias{WKTPlotBM}
\title{Plot Biomass WKT Format function}
\usage{
WKTPlotBM(PlotBio_output, wktfile)
}
\arguments{
\item{PlotBio_output}{- PlotBio Output using PlotIDfile (Should have Lat and Long Columns).}
}
\description{
This function generates a WKTfile using PlotBio function output.
It can be imported into QGIS
}
\details{
It generates a WKTfile for easy loading the biomass files on any GIS platform (e.g. QGIS).
}
\examples{
##Sample data is in data folder
field_data = read.csv("data/sample_tree_data.csv") #read a csv file to a dataframe
pidfile = "data/sample_plotid_file.csv" # path to plotid locations file
samplewktfile = "data/sample_output.wkt"

output = PlotBio(field_data,PlotIDfile=pidfile) # attaches lat and long information

WKTPlotBM(output,wktfile=samplewktfile)

}
\keyword{Biomass}

